TIP-842: Move family model to Akeneo\Pim\Structure
