+++ src/Pim/Bundle/CatalogBundle/spec/Entity/FamilySpec.php
@@ -3,7 +3,7 @@
 namespace spec\Pim\Bundle\CatalogBundle\Entity;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Entity\Family;
+use Akeneo\Pim\Structure\Component\Model\Family;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 
 class FamilySpec extends ObjectBehavior
